requirements.txt
streamlit
PyMuPDF
numpy
pillow
torch
transformers